New command: activate_farm
This clicks on a known safe area, actually 15 pixels above the
farm in the middle right – to the left of the "Help" menu but before
the blue "FarmVille Requests".
This spot could be volatile, but it is a heck of a lot more stable than
any spot within the actual Flash farm area.