This branch holds scripts not yet integrated into the system.
They might be tests (collect_fuel) or scripts that were used
once but deemed to be not fully baked.

Many scripts rely on DSL calls added to lib/farmer_otto.rb, 
which may or may not have been committed. In other words, rebase this branch on top of master frequently.